+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/Cert.hs
@@ -10,7 +10,9 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Cardano.Ledger.Dijkstra.Rules.Cert () where
+module Cardano.Ledger.Dijkstra.Rules.Cert (
+  DijkstraCERT,
+) where
 
 import Cardano.Ledger.BaseTypes (ShelleyBase, StrictMaybe (..))
 import Cardano.Ledger.Conway.Rules (
@@ -25,10 +27,12 @@ import Cardano.Ledger.Conway.Rules (
 import Cardano.Ledger.Conway.TxCert
 import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Era
+import Cardano.Ledger.Dijkstra.Rules.GovCert (DijkstraGovCertPredFailure)
 import Cardano.Ledger.Dijkstra.State
 import Cardano.Ledger.Dijkstra.TxCert
 import Cardano.Ledger.Shelley.Rules (PoolEnv (..), ShelleyPoolPredFailure)
 import Control.State.Transition.Extended
+import Data.Void (absurd)
 import Lens.Micro ((&), (.~), (^.))
 
 type instance EraRuleFailure "CERT" DijkstraEra = ConwayCertPredFailure DijkstraEra
@@ -43,8 +47,11 @@ instance InjectRuleFailure "CERT" ConwayDelegPredFailure DijkstraEra where
 instance InjectRuleFailure "CERT" ShelleyPoolPredFailure DijkstraEra where
   injectFailure = PoolFailure
 
-instance InjectRuleFailure "CERT" ConwayGovCertPredFailure DijkstraEra where
+instance InjectRuleFailure "CERT" DijkstraGovCertPredFailure DijkstraEra where
   injectFailure = GovCertFailure
+
+instance InjectRuleFailure "CERT" ConwayGovCertPredFailure DijkstraEra where
+  injectFailure = GovCertFailure . injectFailure
 
 instance
   ( Era era
@@ -112,3 +119,13 @@ certTransition = do
     DijkstraTxCertGov govCert -> do
       trans @(EraRule "GOVCERT" era) $
         TRC (ConwayGovCertEnv pp currentEpoch committee committeeProposals, certState, govCert)
+
+instance
+  ( Era era
+  , STS (DijkstraGOVCERT era)
+  , PredicateFailure (EraRule "GOVCERT" era) ~ DijkstraGovCertPredFailure era
+  ) =>
+  Embed (DijkstraGOVCERT era) (DijkstraCERT era)
+  where
+  wrapFailed = GovCertFailure
+  wrapEvent = absurd